% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bank.R
\docType{data}
\name{bank}
\alias{bank}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 53940 rows and 10 variables:
\describe{
\item{arrival_time}{price, in US dollars}
\item{service_time}{weight of the diamond, in carats}
...
}
}
\usage{
bank
}
\description{
A dataset containing the prices and other attributes of almost 54,000
banks.
}
\keyword{datasets}
