% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Multiserver.R
\name{Multiserver}
\alias{Multiserver}
\title{Customer Service Times at a Bank}
\usage{
Multiserver(Arrivals, ServiceTimes, NumServers = 1)
}
\arguments{
\item{Arrivals}{text}

\item{ServiceTimes}{text}

\item{NumServers}{text}
}
\value{
dataframe is returned
}
\description{
Customer Service Times at a Bank
}
\examples{
data(bank)
Multiserver(bank$arrival_time, bank$service_time, 2)
}
